Initial code to allow for basic migrations
This variant only includes sparql queries as migrations.  Other
migrations ought to be made possible too.

The migrations currently run when a call is made to the service.  This
should be altered so the service executes the migrations once the
service and the database are up.  Such an approach would allow users to
dump migrations in the folder, restart the migrations service, and stop
caring about things executing correctly.